Added python scripts to generate and clear test files
All test files are in the test directory. makeTestFiles.py generates a
series of files. One of them has every character once, the rest are
weighted to a certain character. The last one is just a bunch of
gibberish. clearTestFiles.py gets rid of the generated files.
runTests.py doesn't do anything at this point.