Added file with RLE methods
This file contains the methods to do the RLE compression and
decompression. Now the methods just need to be worked into the existing
framework. Until we get it integrated you can test it by running ./rle
input_file_name _output_file_name it will automatically compress and
decompress and leaving a compressed version of the file with an .rle
extension.